+++ backend/main.go
@@ -7,11 +7,27 @@ import (
 	"backend/utils"
 	"os"
 
+	log "github.com/sirupsen/logrus"
 	"github.com/subosito/gotenv"
 )
 
 func init() {
 	gotenv.Load()
+
+	// Log as JSON instead of the default ASCII formatter.
+	log.SetFormatter(&log.JSONFormatter{})
+
+	// Output to stdout instead of the default stderr
+	// Can be any io.Writer, see below for File example
+	log.SetOutput(os.Stdout)
+
+	// Only log the warning severity or above.
+	logLevel, err := log.ParseLevel(os.Getenv("LOG_LEVEL"))
+	if err != nil {
+		logLevel = log.InfoLevel
+	}
+
+	log.SetLevel(logLevel)
 }
 
 func main() {